handshake response packet

+++ src/main/java/org/xyz/proxy/net/proto/mysql/MySQLPacket.java
@@ -121,7 +121,7 @@ public abstract class MySQLPacket {
     public static final byte COM_BINLOG_DUMP = 0x12;
 
     private int payloadLength;
-    private byte sequenceId = 0;
+    private int sequenceId = 0;
 
     /**
      * 把数据包写到buffer中，如果buffer满了就把buffer通过前端连接写出。
